requirements.txt
alembic==1.13.1
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docker==7.0.0
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.2
importlib-resources==6.1.3
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.5
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mlflow==2.11.1
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.10.2
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic-core==2.16.3
pydeck==0.8.1b0
PyDrive==1.3.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
streamlit==1.32.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==1.10.0
torchvision==0.11.1
tornado==6.4
traitlets==5.14.2
triton==2.2.0
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.41.2
yarg==0.1.9
zipp==3.17.0
